+++ mods/BrewLAN/documentation/Changelog.md
@@ -981,3 +981,7 @@ All changes by Sean Wheeldon (Balthazar) unless otherwise stated.
   ### 2021-08-09
 
   * Replaced the `SHIELD` category with `PERSONALSHIELD` on the Aeon shielded resource structures.
+
+  ### 2021-08-12
+
+  * Caffee Corretto: Tesla Coil now doesn't get damage reductions from effects that reduce its maintenance cost. It, however, still gets damage increases from effects that increase the maintenance cost.